third_party/blink/web_tests/mhtml/page_with_css_and_js_ie-expected.txt

CONSOLE ERROR: Blocked script execution in 'page_with_css_and_js_ie.mht' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE ERROR: Blocked script execution in 'page_with_css_and_js_ie.mht' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
